Implemented search based on Trip Type and added few test cases

+++ backend/flight-service/flight-service/src/main/java/com/flightapp/flightservice/dto/FlightResponse.java
@@ -1,5 +1,6 @@
 package com.flightapp.flightservice.dto;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 import lombok.AllArgsConstructor;
@@ -24,6 +25,6 @@ public class FlightResponse {
     private LocalDateTime arrivalTime;
     private Integer economySeats;
     private Integer businessSeats;
-    private float ticketCost;
+    private BigDecimal ticketCost;
     private String status;
 }
