+++ backend/flight-service/flight-service/src/main/java/com/flightapp/flightservice/model/Flight.java
@@ -3,8 +3,21 @@ package com.flightapp.flightservice.model;
 import com.flightapp.flightservice.enums.FlightStatus;
 import com.flightapp.flightservice.enums.MealType;
 
-import jakarta.persistence.*;
-import lombok.*;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.PrePersist;
+import jakarta.persistence.PreUpdate;
+import jakarta.persistence.Table;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
 
 import java.math.BigDecimal;
 import java.time.LocalDateTime;
@@ -37,7 +50,7 @@ public class Flight {
     @Column(nullable = false)
     private Integer businessSeats;
     @Column(nullable = false)
-    private float ticketCost;
+    private BigDecimal ticketCost;
     @Enumerated(EnumType.STRING)
     @Column(nullable = false,length=20)
     private MealType mealType;
